build: Move installation of cargo2junit to builder (#70)
* build: Move installation of cargo2junit to builder

Install nigthly and use that to install cargo2junit.  This avoids constant installations for each test run, as well as moving the details about the required toolchain for cargo2junit out of the build scripts.

* build: Switch to comet-builder

Originally the comet-builder would fail because of XML comments in one of the toml files.  This has been addressed.