feat: Optimze CreateNamedStruct preserve dictionaries (#789)
* feat: Optimze CreateNamedStruct preserve dictionaries

Instead of serializing the return data_type we just serialize the field
names. The original implmentation was done as it lead to slightly
simpler implementation, but it clear from #750 that this was the wrong
choice and leads to issues with the physical data_type.

* Support dictionary data_types in StructVector and MapVector

* Add length checks

(cherry picked from commit e8d94d69ec0a59fa369dd54198d8cda74186555f)