Make it possible to run 'make benchmark-%' using jvm 17+ (#823)
When using jvm 17+ spark needs extra jvm args to avoid getting errors
like
```
[ERROR] Failed to execute goal org.codehaus.mojo:exec-maven-plugin:3.2.0:java (default-cli) on project comet-spark-spark3.4_2.12: An exception occurred while executing the Java class. class org.apache.spark.storage.StorageUtils$ (in unnamed module @0x35b75242) cannot access class sun.nio.ch.DirectBuffer (in module java.base) because module java.base does not export sun.nio.ch to unnamed module @0x35b75242 -> [Help 1]
```

These args are already present inside the main pom.xml. To avoid
duplicating the args we using maven to extract them. In order avoid
slowing down the Makefile the args are defined as a function and
therefore only evaluated when they are needed.