Make Comet opt-in but with all features enabled by default (#557)
## Which issue does this PR close?

N/A

## Rationale for this change

We know that users often need to change Spark's memory configuration
when enabling Comet and that there are limitations when using on-heap
memory.

We also know that some jobs can fail with OOM when performing Sort,
which is a very common operation because it is used for most joins,
since Spark tends to use SortMergeJoin. This is a current limitation in
DataFusion. See the following issues for more details:

- https://github.com/apache/datafusion/issues/14692
- https://github.com/apache/datafusion/issues/13431
- https://github.com/apache/datafusion/issues/15028

## What changes are included in this PR?

- Disable Comet by default, but user can opt-in
- Ensure that Comet is enabled when running tests in CI

## How are these changes tested?