feat: make parquet native scan schema case insensitive (#1575)
## Which issue does this PR close?

Part of #1574.

## Rationale for this change

The data schema of spark parquet datasource scan may not be consistent with the actual schema of file. This may cause the `projection/filter` to not behave as expected.

## What changes are included in this PR?

This PR makes `data schema` and `file schema` case insensitive in schema adapter, but it does not affect `pruning_predicate` and `page_pruning_predicate`.

## How are these changes tested?

added unit test