Run PRB-like builds in main-apple (#555)
## Which issue does this PR close?

We want to run the same tests in our `main-apple` branch `build` runs as
we do with `prb` runs.

It is intended to be used for both our development builds (`main-apple`)
as well as branch-specific builds (`branch-0.7-apple`). Due to the way
that Rio works, we'll still need to adjust the build because
`pipelineChain.pipeline` doesn't support variable substitution.

## Rationale for this change

As we move to automatically apply upstream OSS changes, we want to
re-run the tests to verify that they haven't introduced an issue.

## What changes are included in this PR?

Use a branch rule to test in both `main-apple` and `branch-.*-apple`
branches, applying the same tests as we run in `prb` jobs.

Rename a job which has a corrupted workspace.
rdar://148542912 (ARM builds are failing because it can't extract the workspace from the cache)

## How are these changes tested?

The jobs are exactly the same using Rio branch rules and YML references.
Verify that the references can be processed by using `yq`.