build: Automate sync on feature branch (#575)
## Which issue does this PR close?

This runs a cherry-picking script which automates applying OSS changes.

## Rationale for this change

The team should see merge and benchmark upstream changes to catch issues
as early as possible.

## What changes are included in this PR?

Adds `dev/cometSync.sh` script and hourly Rio job to pick up new changes
from upstream. It uses a tracking tag `MERGED_<branch>` to identify the
last merged change.

The `.cicd` branch was added to the ignore list since this is used by
the Rio build system.

## How are these changes tested?

A branch `feature-sync-apple` was used in combination with a small
temporary change that always chose the `main` branch (instead of
`feature-sync`) to demonstrate that changes were automatically applied.